+++ src/util/mod.rs
@@ -44,7 +44,6 @@ pub mod file;
 pub mod file_log;
 pub mod clocktime;
 pub mod metrics;
-pub mod profiling;
 #[cfg(target_os="linux")]
 mod thread_metrics;
